Feature/evaluated

+++ src/main/java/com/square/Inventory/Management/System/Repository/InvoicesRepository.java
@@ -3,5 +3,5 @@ package com.square.Inventory.Management.System.Repository;
 import com.square.Inventory.Management.System.Entity.Invoices;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoicesRepository extends JpaRepository<Invoices,Integer> {
+public interface InvoicesRepository extends JpaRepository<Invoices,Long> {
 }
